requirements.txt
streamlit>=1.32
plotly>=5.18
pandas>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0